requirements.txt
aiolimiter>=1.1.0
duckdb>=0.10.3
httpx>=0.27.0
pandas>=2.2.0
pandas_market_calendars>=4.4.0
pyyaml>=6.0.1
streamlit>=1.37.0
pytest>=8.0.0
